feature: post metadata migrations updated

+++ migrations/2025-03-02-065637_posts_metadata/up.sql
@@ -1,7 +1,7 @@
 ALTER TABLE public.posts
-    ADD COLUMN post_view_count integer DEFAULT 0 NOT NULL,
-    ADD COLUMN post_upvote_count integer DEFAULT 0 NOT NULL,
-    ADD COLUMN post_share_count integer DEFAULT 0 NOT NULL,
+    ADD COLUMN post_view_count int8 DEFAULT 0 NOT NULL,
+    ADD COLUMN post_upvote_count int8 DEFAULT 0 NOT NULL,
+    ADD COLUMN post_share_count int8 DEFAULT 0 NOT NULL,
     ADD COLUMN post_metadata jsonb DEFAULT '{}'::jsonb NOT NULL;
 
 -- Add indexes for efficient querying of popular posts
